Refactor test location so it matches astronomer/astronomer
