+++ .pre-commit-config.yaml
@@ -35,12 +35,10 @@ repos:
         args: ["-b", "master"]
       - id: requirements-txt-fixer
         args:
-          [
-            "tests/chart_tests/requirements.in",
-            "tests/chart_tests/requirements.txt",
-            "tests/functional-tests/requirements.in",
-            "tests/functional-tests/requirements.txt",
-          ]
+          - "requirements/chart.in"
+          - "requirements/chart.txt"
+          - "requirements/functional.in"
+          - "requirements/functional.txt"
       - id: sort-simple-yaml
       - id: trailing-whitespace
   - repo: https://github.com/shellcheck-py/shellcheck-py
