+++ bin/run-ci
@@ -142,7 +142,7 @@ echo "Airflow Version: $AIRFLOW_VERSION"
 pip3 install virtualenv
 virtualenv --python=python3 /tmp/venv
 source /tmp/venv/bin/activate
-pip install -r "$REPO_DIR/tests/functional-tests/requirements.txt"
+pip install -r "$REPO_DIR/requirements/functional.txt"
 mkdir -p "test-results"
-pytest -v "--junitxml=test-results/junit.xml" "$REPO_DIR/tests/functional-tests/"
+pytest -v "--junitxml=test-results/junit.xml" "$REPO_DIR/tests/functional/"
 deactivate
