+++ tests/chart/test_dag_server_serviceaccount.py
@@ -1,6 +1,6 @@
 import pytest
 
-from tests.chart_tests.helm_template_generator import render_chart
+from tests.chart.helm_template_generator import render_chart
 
 from .. import supported_k8s_versions
