+++ tests/chart/test_extra_objects.py
@@ -3,7 +3,7 @@ from subprocess import CalledProcessError
 
 import pytest
 
-from tests.chart_tests.helm_template_generator import render_chart
+from tests.chart.helm_template_generator import render_chart
 
 from .. import supported_k8s_versions
