+++ tests/chart/test_logging_sidecar.py
@@ -3,7 +3,7 @@ import textwrap
 import pytest
 import yaml
 
-from tests.chart_tests.helm_template_generator import render_chart
+from tests.chart.helm_template_generator import render_chart
 
 from .. import supported_k8s_versions
