+++ tests/chart/test_pgbouncer.py
@@ -1,6 +1,6 @@
 import base64
 
-from tests.chart_tests.helm_template_generator import render_chart
+from tests.chart.helm_template_generator import render_chart
 
 
 class TestPgbouncerSecret:
