+++ tests/chart/test_pgbouncer_certgenerator.py
@@ -1,7 +1,7 @@
 import jmespath
 import pytest
 
-from tests.chart_tests.helm_template_generator import render_chart
+from tests.chart.helm_template_generator import render_chart
 
 from .. import supported_k8s_versions
